Add another example
