+++ kafka-worker.gemspec
@@ -12,5 +12,4 @@ Gem::Specification.new do |s|
   s.homepage     = 'https://github.com/creatubbles/ruby-kafka-worker'
 
   s.add_dependency('ruby-kafka')
-  s.add_dependency('activesupport')
 end
