+++ lib/kafka-worker.rb
@@ -1,4 +1,3 @@
-require 'active_support/concern'
 require 'kafka'
 
 module KafkaWorker
@@ -91,81 +90,3 @@ module KafkaWorker
     end
   end
 end
-
-
-=begin
-# created only once
-class ClassAbstractCtbHandler
-  cattr_accessor :ctb_client
-end
-
-# created each time useful if you need to mutate
-class AbstractCtbHandler  
-  attr_accessor :ctb_client
-  def initialize
-    @ctb_client = Ctb::Creatubbles.new(
-      client_id: ENV['CREATUBBLES_CLIENT_ID'],
-      client_secret: ENV['CREATUBBLES_CLIENT_SECRET'],
-      api_url: ENV['CREATUBBLES_API_URL'] || Creatubbles::DEFAULT_API_URL 
-    )
-  end
-end
-
-class Handler < AbstractCtbHandler
-  include KafkaWorker::Handler
-  consumes "hello"
-
-  def handle(message)
-    puts 'handle hello'
-    puts message.inspect
-    ctb_client
-  end
-end
-
-class SecondHandler < ClassAbstractCtbHandler
-  include KafkaWorker::Handler
-  consumes "goodbye"
-
-  def handle(message)
-    puts 'handle goodbye'
-    puts message.inspect
-    ctb_client
-  end
-end
-
-ClassAbstractCtbHandler.ctb_client = Ctb::Creatubbles.new(
-  client_id: ENV['CREATUBBLES_CLIENT_ID'],
-  client_secret: ENV['CREATUBBLES_CLIENT_SECRET'],
-  api_url: ENV['CREATUBBLES_API_URL'] || Creatubbles::DEFAULT_API_URL 
-)
-
-opts = {
-#   client_id: ENV['CREATUBBLES_CLIENT_ID'],
-#   client_secret: ENV['CREATUBBLES_CLIENT_SECRET'],
-#   api_url: ENV['CREATUBBLES_API_URL'] || Creatubbles::DEFAULT_API_URL,
-   kafka_ips: "127.0.0.1:9092",
-   client_id: 'test',
-   group_id: 'test'
-}
-kw = KafkaWorker::Worker.new(opts)
-kw.run
-trap("QUIT") { kw.stop_consumer }
-=end
-
-
-=begin
-opts = {
-   client_id: ENV['CREATUBBLES_CLIENT_ID'],
-   client_secret: ENV['CREATUBBLES_CLIENT_SECRET'],
-   api_url: ENV['CREATUBBLES_API_URL'] || Creatubbles::DEFAULT_API_URL,
-   kafka_ips: (ENV['KAFKA_ENDPOINTS'] || "localhost").split(",")
-}
-=end
-
-=begin
-opts = {
-   kafka_ips: (ENV['KAFKA_ENDPOINTS'] || "localhost").split(",")
-}
-k = KafkaWorker::Worker.new(opts)
-k.run
-=end
